Fix ManagedUpload never finishes when stream size == part size
When uploading a stream with an unknown initial length and a total length
exactly equal to a multiple of the part size, the upload callback is
never finished.

This is caused by only checking whether the upload finished after uploading
the last part, but not when the end event is received.

In our case the end event is received after the last part has been uploaded,
but there is also no more data to be uploaded, resulting in a hanging upload.